pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crownan"
dynamic = ["version"]
description = "AI-powered tools for the Krónan supermarket API"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [{ name = "Kjartan Palsson" }]
keywords = ["kronan", "grocery", "iceland", "api", "ai", "slack"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
agent = [
    "anthropic>=0.52.0",
]
slackbot = [
    "anthropic>=0.52.0",
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.27.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[project.scripts]
crownan = "crownan.cli:main"
crownan-agent-setup = "crownan.agent.setup:main"
crownan-benchmark = "crownan.benchmark:main"

[project.urls]
Homepage = "https://github.com/kjarripa/crownan"
Repository = "https://github.com/kjarripa/crownan"
Issues = "https://github.com/kjarripa/crownan/issues"

[tool.hatch.version]
path = "src/crownan/_version.py"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.lint.per-file-ignores]
# Tool definitions have long description strings (sent to API as-is)
"src/crownan/agent/tools.py" = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
